splits breadcrumb component into many
