+++ packages/manager/src/components/Breadcrumb/types.ts
@@ -1,0 +1,17 @@
+import { CSSProperties } from 'src/components/core/styles';
+
+export interface LabelProps {
+  linkTo?: string;
+  prefixComponent?: JSX.Element | null;
+  prefixStyle?: CSSProperties;
+  suffixComponent?: JSX.Element | null;
+  subtitle?: string;
+  noCap?: boolean;
+}
+
+export interface EditableProps {
+  onCancel: () => void;
+  onEdit: (value: string) => Promise<any>;
+  errorText?: string;
+  editableTextTitle: string;
+}
